+++ lib/powerpack/array.rb
@@ -1,2 +1,3 @@
 require_relative 'array/butlast'
 require_relative 'array/butfirst'
+require_relative 'array/unwrap'